+++ backend/src/modules/core/container.ts
@@ -23,6 +23,7 @@ import { QueueService } from '../question/services/QueueService.js';
 import { RoleAssigneeService } from '../question/services/RoleAssigneeService.js';
 import { AllocationService } from '../question/services/AllocationService.js';
 import { ModeratorQueueService } from '../question/services/ModeratorQueueService.js';
+import { QuestionMaintenanceService } from '../question/services/QuestionMaintenanceService.js';
 import { ContextController } from '../context/controllers/ContextController.js';
 import { ContextService } from '../context/services/ContextService.js';
 import { PerformanceController } from '../performance/controllers/PerformanceController.js';
@@ -85,6 +86,10 @@ export const coreContainerModule = new ContainerModule(options => {
   options
     .bind(GLOBAL_TYPES.ModeratorQueueService)
     .to(ModeratorQueueService)
+    .inSingletonScope();
+  options
+    .bind(GLOBAL_TYPES.QuestionMaintenanceService)
+    .to(QuestionMaintenanceService)
     .inSingletonScope();
   options.bind(CORE_TYPES.AnswerService).to(AnswerService).inSingletonScope();
   options.bind(CORE_TYPES.ContextService).to(ContextService).inSingletonScope();
