+++ backend/src/modules/question/services/QuestionService.ts
@@ -8,6 +8,7 @@ import { IQueueService } from '../interfaces/IQueueService.js';
 import { IRoleAssigneeService } from '../interfaces/IRoleAssigneeService.js';
 import { IAllocationService } from '../interfaces/IAllocationService.js';
 import { IModeratorQueueService } from '../interfaces/IModeratorQueueService.js';
+import { IQuestionMaintenanceService } from '../interfaces/IQuestionMaintenanceService.js';
 import { resolveExpertMeta } from './helpers/reportHelpers.js';
 import { BaseService, MongoDatabase } from '#root/shared/index.js';
 import { GLOBAL_TYPES } from '#root/types.js';
@@ -18,7 +19,6 @@ import {
   IQuestion,
   IUser,
   IQuestionSubmission,
-  ISubmissionHistory,
   IAnswer,
   INotificationType,
   IQuestionPriority,
@@ -43,7 +43,6 @@ import { IQuestionSubmissionRepository } from '#root/shared/database/interfaces/
 import { IUserRepository } from '#root/shared/database/interfaces/IUserRepository.js';
 import { IRequestRepository } from '#root/shared/database/interfaces/IRequestRepository.js';
 import { IContextRepository } from '#root/shared/database/interfaces/IContextRepository.js';
-import { INotificationRepository } from '#root/shared/database/interfaces/INotificationRepository.js';
 import { notifyUser } from '#root/utils/pushNotification.js';
 import { normalizeKeysToLower } from '#root/utils/normalizeKeysToLower.js';
 import { appConfig } from '#root/config/app.js';
@@ -122,9 +121,6 @@ export class QuestionService extends BaseService implements IQuestionService {
     @inject(GLOBAL_TYPES.AnswerRepository)
     private readonly answerRepo: IAnswerRepository,
 
-    @inject(GLOBAL_TYPES.NotificationRepository)
-    private readonly notificationRepository: INotificationRepository,
-
     @inject(GLOBAL_TYPES.NotificationService)
     private readonly notificationService: NotificationService,
 
@@ -178,6 +174,9 @@ export class QuestionService extends BaseService implements IQuestionService {
 
     @inject(GLOBAL_TYPES.ModeratorQueueService)
     private readonly moderatorQueueService: IModeratorQueueService,
+
+    @inject(GLOBAL_TYPES.QuestionMaintenanceService)
+    private readonly maintenanceService: IQuestionMaintenanceService,
   ) {
     super(mongoDatabase);
   }
@@ -371,63 +370,17 @@ export class QuestionService extends BaseService implements IQuestionService {
   /** Standardise a state name across all questions: any question whose details.state matches
    *  one of `currentValues` (e.g. "punjab", "PUNJAB", "पंजाब") is set to `standardizedTo`
    *  (e.g. "Punjab"). Returns how many matched/were modified. */
-  async normalizeQuestionState(
-    currentValues: string[],
-    standardizedTo: string,
-  ): Promise<{matched: number; modified: number}> {
-    const cleaned = (currentValues ?? [])
-      .map(v => (typeof v === 'string' ? v.trim() : ''))
-      .filter(Boolean);
-    const target = (standardizedTo ?? '').trim();
-    if (cleaned.length === 0) {
-      throw new BadRequestError(
-        'current values must be a non-empty array of strings',
-      );
-    }
-    if (!target) {
-      throw new BadRequestError('standardizedTo is required');
-    }
-    return this._withTransaction(async (session: ClientSession) => {
-      return this.questionRepo.normalizeQuestionState(cleaned, target, session);
-    });
+  // ── Admin / maintenance / normalization delegates to QuestionMaintenanceService ──
+  async normalizeQuestionState(currentValues: string[], standardizedTo: string) {
+    return this.maintenanceService.normalizeQuestionState(currentValues, standardizedTo);
   }
 
-  /** Standardise question district names, validating each `standardiseTo` against the
-   *  `districts` collection (districtNameEnglish). Matching ones update questions whose
-   *  details.district === existingName; non-matching names are returned untouched. */
-  async normalizeQuestionDistricts(
-    mappings: {existingName: string; standardiseTo: string}[],
-  ) {
-    const cleaned = (mappings ?? [])
-      .map(m => ({
-        existingName:
-          typeof m?.existingName === 'string' ? m.existingName.trim() : '',
-        standardiseTo:
-          typeof m?.standardiseTo === 'string' ? m.standardiseTo.trim() : '',
-      }))
-      .filter(m => m.existingName && m.standardiseTo);
-    if (cleaned.length === 0) {
-      throw new BadRequestError(
-        'mappings must be a non-empty array of { existingName, standardiseTo }',
-      );
-    }
-    return this.questionRepo.normalizeQuestionDistricts(cleaned);
+  async normalizeQuestionDistricts(mappings: {existingName: string; standardiseTo: string}[]) {
+    return this.maintenanceService.normalizeQuestionDistricts(mappings);
   }
 
-  /** Audit: distinct question details.state / details.district values that don't exist in the
-   *  states / districts collections. */
-  async findUnknownQuestionGeo(): Promise<{
-    unknownStates: string[];
-    matchedDistricts: {
-      name: string;
-      foundIn: 'block' | 'village';
-      districtCode: number | null;
-      stateCode: number | null;
-      districtNameEnglish: string | null;
-    }[];
-    notMatchingDistricts: string[];
-  }> {
-    return this.questionRepo.findUnknownQuestionGeo();
+  async findUnknownQuestionGeo() {
+    return this.maintenanceService.findUnknownQuestionGeo();
   }
 
   async getAllocatedQuestions(
@@ -2635,97 +2588,12 @@ export class QuestionService extends BaseService implements IQuestionService {
   }
 
   //send notification to moderators for delayed questions
-  async sendDelayedNotifications(): Promise<void> {
-    await this._withTransaction(async session => {
-      const delayedReviews =
-        await this.questionSubmissionRepo.getDelayedReviews(session);
-      if (!delayedReviews.length) {
-        return;
-      }
-
-      const notifiedSubmissionIds: ObjectId[] = [];
-
-      const moderators = await this.userRepo.findModerators();
-
-      for (const item of delayedReviews) {
-        try {
-          await Promise.allSettled(
-            moderators.map(mod =>
-              this.notificationRepository.addNotification(
-                mod._id.toString(),
-                item.questionId.toString(),
-                'question_delayed',
-                'A question has been delayed for 45 minutes',
-                'Question Delayed',
-              ),
-            ),
-          );
-
-          notifiedSubmissionIds.push(item?._id);
-        } catch (error) {
-          console.error(
-            `Failed notification for question ${item?.questionId}`,
-            error,
-          );
-        }
-      }
-      if (notifiedSubmissionIds.length) {
-        await this.questionSubmissionRepo.markDelayedNotificationsSent(
-          notifiedSubmissionIds,
-          session,
-        );
-      }
-    });
+  async sendDelayedNotifications() {
+    return this.maintenanceService.sendDelayedNotifications();
   }
 
-  async backfillEmptyEmbeddings(batchLimit = 50): Promise<void> {
-    if (!appConfig.ENABLE_AI_SERVER) {
-      console.log('<<EMBEDDING_BACKFILL>> AI server disabled, skipping.');
-      return;
-    }
-
-    const questions =
-      await this.questionRepo.getQuestionsWithEmptyEmbeddings(batchLimit);
-
-    if (questions.length === 0) {
-      console.log(
-        '<<EMBEDDING_BACKFILL>> No questions with empty embeddings found.',
-      );
-      return;
-    }
-
-    console.log(
-      `<<EMBEDDING_BACKFILL>> Processing ${questions.length} question(s)...`,
-    );
-
-    let succeeded = 0;
-    let failed = 0;
-
-    for (const q of questions) {
-      const inputText = (q.question || q.text || '').trim();
-
-      if (!inputText) {
-        console.warn(`<<EMBEDDING_BACKFILL>> Skipping ${q._id} — no text`);
-        failed++;
-        continue;
-      }
-
-      try {
-        const {embedding} = await this.aiService.getEmbedding(inputText);
-        await this.questionRepo.updateQuestionEmbedding(
-          q._id.toString(),
-          embedding,
-        );
-        succeeded++;
-      } catch (err) {
-        console.error(`<<EMBEDDING_BACKFILL>> Failed for ${q._id}:`, err);
-        failed++;
-      }
-    }
-
-    console.log(
-      `<<EMBEDDING_BACKFILL>> Done — ✅ ${succeeded} succeeded, ❌ ${failed} failed`,
-    );
+  async backfillEmptyEmbeddings(batchLimit?: number) {
+    return this.maintenanceService.backfillEmptyEmbeddings(batchLimit);
   }
 
   // ─── Time-bound question tracking ───────────────────────────────────────────
@@ -2830,120 +2698,24 @@ export class QuestionService extends BaseService implements IQuestionService {
    * This is used for migration purposes to fix duplicate entries.
    * @param submissionId - The submission document ID
    */
-  async backgroundProcessAction(
-    userId: string,
-  ): Promise<{modifiedCount: number}> {
-    return await this.userRepo.clearAssignedQuestions(userId);
+  async backgroundProcessAction(userId: string) {
+    return this.maintenanceService.backgroundProcessAction(userId);
   }
 
-  /** Admin utility: remove a submission history entry (by 0-based index) for a question. */
-  async removeSubmissionHistoryEntry(
-    questionId: string,
-    index: number,
-  ): Promise<{success: boolean; historyLength: number}> {
-    const updated = await this.questionSubmissionRepo.removeHistoryEntryByIndex(
-      questionId,
-      index,
-    );
-    return {
-      success: true,
-      historyLength: updated?.history?.length ?? 0,
-    };
+  async removeSubmissionHistoryEntry(questionId: string, index: number) {
+    return this.maintenanceService.removeSubmissionHistoryEntry(questionId, index);
   }
 
-  /** Admin data-fix: remove a single expert from a question's submission queue by index. */
-  async removeSubmissionQueueEntry(
-    questionId: string,
-    index: number,
-  ): Promise<{success: boolean; queueLength: number}> {
-    const updated = await this.questionSubmissionRepo.removeQueueEntryByIndex(
-      questionId,
-      index,
-    );
-    return {
-      success: true,
-      queueLength: updated?.queue?.length ?? 0,
-    };
+  async removeSubmissionQueueEntry(questionId: string, index: number) {
+    return this.maintenanceService.removeSubmissionQueueEntry(questionId, index);
   }
 
-  /** Admin utility: append an expert to a question's submission queue. */
-  async addSubmissionQueueEntry(
-    questionId: string,
-    expertId: string,
-  ): Promise<{success: boolean; queueLength: number}> {
-    if (!expertId || !ObjectId.isValid(expertId)) {
-      throw new BadRequestError('A valid expertId is required');
-    }
-    const updated = await this.questionSubmissionRepo.addQueueEntry(
-      questionId,
-      expertId,
-    );
-    return {success: true, queueLength: updated?.queue?.length ?? 0};
+  async addSubmissionQueueEntry(questionId: string, expertId: string) {
+    return this.maintenanceService.addSubmissionQueueEntry(questionId, expertId);
   }
 
-  /** Admin utility: append a history entry to a question's submission history.
-   *  The raw entry's id/date fields are coerced to ObjectId/Date before storing. */
-  async addSubmissionHistoryEntry(
-    questionId: string,
-    rawEntry: Record<string, any>,
-  ): Promise<{success: boolean; historyLength: number}> {
-    const entry = this.buildHistoryEntry(rawEntry);
-    const updated = await this.questionSubmissionRepo.addHistoryEntry(
-      questionId,
-      entry,
-    );
-    return {
-      success: true,
-      historyLength: updated?.history?.length ?? 0,
-    };
-  }
-
-  /** Coerce a raw JSON history entry into a stored ISubmissionHistory (ObjectIds + Dates). */
-  private buildHistoryEntry(raw: Record<string, any>): ISubmissionHistory {
-    if (!raw || typeof raw !== 'object') {
-      throw new BadRequestError('entry object is required');
-    }
-    const toOid = (v: unknown): ObjectId => {
-      if (!v || !ObjectId.isValid(String(v))) {
-        throw new BadRequestError(`Invalid ObjectId: ${String(v)}`);
-      }
-      return new ObjectId(String(v));
-    };
-    const toDate = (v: unknown): Date | undefined =>
-      v ? new Date(v as string) : undefined;
-
-    if (!raw.updatedBy) {
-      throw new BadRequestError('entry.updatedBy is required');
-    }
-    if (!raw.status) {
-      throw new BadRequestError('entry.status is required');
-    }
-
-    const entry: any = {
-      updatedBy: toOid(raw.updatedBy),
-      status: raw.status,
-      createdAt: toDate(raw.createdAt) ?? new Date(),
-      updatedAt: toDate(raw.updatedAt) ?? new Date(),
-    };
-
-    // Optional ObjectId fields.
-    if (raw.answer) entry.answer = toOid(raw.answer);
-    if (raw.reviewId) entry.reviewId = toOid(raw.reviewId);
-    if (raw.approvedAnswer) entry.approvedAnswer = toOid(raw.approvedAnswer);
-    if (raw.rejectedBy) entry.rejectedBy = toOid(raw.rejectedBy);
-    if (raw.rejectedAnswer) entry.rejectedAnswer = toOid(raw.rejectedAnswer);
-    if (raw.lastModifiedBy) entry.lastModifiedBy = toOid(raw.lastModifiedBy);
-    if (raw.modifiedAnswer) entry.modifiedAnswer = toOid(raw.modifiedAnswer);
-
-    // Optional string fields.
-    if (raw.reasonForRejection) {
-      entry.reasonForRejection = String(raw.reasonForRejection);
-    }
-    if (raw.reasonForLastModification) {
-      entry.reasonForLastModification = String(raw.reasonForLastModification);
-    }
-
-    return entry as ISubmissionHistory;
+  async addSubmissionHistoryEntry(questionId: string, rawEntry: Record<string, any>) {
+    return this.maintenanceService.addSubmissionHistoryEntry(questionId, rawEntry);
   }
 
   // Feedback allocation
@@ -3026,103 +2798,16 @@ export class QuestionService extends BaseService implements IQuestionService {
   /** Bulk-set `normalizedDomain` on questions from a list of
    *  { "Question ID", "Standardized Domain" } entries. Returns modified / not-matched
    *  counts. */
-  async setNormalizedDomains(
-    entries: { 'Question ID'?: string; 'Standardized Domain'?: string }[],
-  ): Promise<{
-    total: number;
-    matched: number;
-    modified: number;
-    notMatched: number;
-    invalid: number;
-  }> {
-    const pairs = (Array.isArray(entries) ? entries : []).map(e => ({
-      questionId: String(e?.['Question ID'] ?? '').trim(),
-      normalizedDomain: String(e?.['Standardized Domain'] ?? '').trim(),
-    }));
-    return this.questionRepo.bulkSetNormalizedDomain(pairs);
+  async setNormalizedDomains(entries: {'Question ID'?: string; 'Standardized Domain'?: string}[]) {
+    return this.maintenanceService.setNormalizedDomains(entries);
   }
 
-  /** Diagnostic: closed questions in a window that don't have a final answer with a
-   *  valid ObjectId approvedBy — i.e. the ones dropped from the moderator breakdown,
-   *  which explains the "closed count vs breakdown count" mismatch. */
-  async getClosedAnswerMismatch(
-    startTime?: Date,
-    endTime?: Date,
-  ): Promise<{
-    window: { start: Date; end: Date };
-    totalClosed: number;
-    matched: number;
-    mismatched: number;
-    items: any[];
-  }> {
-    let start = startTime;
-    let end = endTime;
-    if (!start || !end) {
-      // Default to the current IST day if no window is given.
-      start = new Date();
-      start.setHours(0, 0, 0, 0);
-      end = new Date(start);
-      end.setDate(end.getDate() + 1);
-    }
-    return this.questionRepo.getClosedAnswerMismatch(start, end);
+  async getClosedAnswerMismatch(startTime?: Date, endTime?: Date) {
+    return this.maintenanceService.getClosedAnswerMismatch(startTime, endTime);
   }
 
-  async backfillClosedModeratorIds(limit = 500): Promise<{
-    matched: number;
-    updated: number;
-    skippedNoFinalAnswer: number;
-    skippedNoApprover: number;
-  }> {
-    const questionIds =
-      await this.questionRepo.findClosedQuestionsWithoutModerator(limit);
-    if (!questionIds.length) {
-      return {
-        matched: 0,
-        updated: 0,
-        skippedNoFinalAnswer: 0,
-        skippedNoApprover: 0,
-      };
-    }
-
-    const finalAnswers =
-      await this.answerRepo.getFinalAnswersByQuestionIds(questionIds);
-
-    // questionId -> approvedBy (first final answer that has an approver wins).
-    const approverByQuestion = new Map<string, string>();
-    const questionsWithFinal = new Set<string>();
-    for (const a of finalAnswers) {
-      const qid = a.questionId?.toString();
-      if (!qid) continue;
-      questionsWithFinal.add(qid);
-      const approver = a.approvedBy?.toString();
-      if (approver && !approverByQuestion.has(qid)) {
-        approverByQuestion.set(qid, approver);
-      }
-    }
-
-    const pairs: {questionId: string; moderatorId: string}[] = [];
-    let skippedNoFinalAnswer = 0;
-    let skippedNoApprover = 0;
-    for (const qid of questionIds) {
-      if (!questionsWithFinal.has(qid)) {
-        skippedNoFinalAnswer++;
-        continue;
-      }
-      const approver = approverByQuestion.get(qid);
-      if (!approver) {
-        skippedNoApprover++;
-        continue;
-      }
-      pairs.push({questionId: qid, moderatorId: approver});
-    }
-
-    const updated = await this.questionRepo.bulkSetModeratorId(pairs);
-    return {
-      matched: questionIds.length,
-      updated,
-      skippedNoFinalAnswer,
-      skippedNoApprover,
-    };
+  async backfillClosedModeratorIds(limit?: number) {
+    return this.maintenanceService.backfillClosedModeratorIds(limit);
   }
 
   /**
